commit de várias modificações pendentes.

+++ acmus-project/src/acmus/tools/structures/Vector.java
@@ -11,15 +11,15 @@ public class Vector {
 		z = c;
 	}
 	
-	public double getX() {
+	public float getX() {
 		return x;
 	}
 
-	public double getY() {
+	public float getY() {
 		return y;
 	}
 
-	public double getZ() {
+	public float getZ() {
 		return z;
 	}
